+++ dymension/libs/kaspa/demo/relayer/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 core = { path = "../../lib/core" }
 hardcode = { path = "../../lib/hardcode" }
 relayer = { path = "../../lib/relayer" }
-validator = { path = "../../lib/validator" } # TODO: remove
+validator = { path = "../../lib/validator" } #TODO: remove
 kaspa-addresses = { workspace = true }
 kaspa-consensus-core = { workspace = true }
 kaspa-core = { workspace = true }
