+++ dymension/libs/kaspa/lib/hardcode/Cargo.toml
@@ -6,3 +6,7 @@ edition = "2021"
 [dependencies]
 kaspa-addresses = { workspace = true }
 kaspa-consensus-core = { workspace = true }
+api-rs = { path = "../../lib/api", package = "openapi" }
+
+reqwest-middleware = "0.4"
+reqwest = { workspace = true }
