+++ dymension/libs/kaspa/lib/hardcode/src/e2e.rs
@@ -1,7 +1,11 @@
 /* For e2e tests and testing only */
 
+use api_rs::apis::configuration;
 use kaspa_addresses::{Prefix, Version};
 use kaspa_consensus_core::network::{NetworkId, NetworkType};
+
+use reqwest::Client;
+use reqwest_middleware::ClientBuilder;
 
 pub const NETWORK: NetworkType = NetworkType::Testnet;
 pub const NETWORK_ID: NetworkId = NetworkId::with_suffix(NETWORK, 10);
@@ -14,3 +18,19 @@ pub const DEPOSIT_AMOUNT: u64 = 20_000_001;
 
 // How much relayer spends to deliver the tx to the network
 pub const RELAYER_NETWORK_FEE: u64 = 5000;
+
+// TODO: remove
+pub const ESCROW_ADDRESS: &str =
+    "kaspatest:qzwyrgapjnhtjqkxdrmp7fpm3yddw296v2ajv9nmgmw5k3z0r38guevxyk7j0";
+
+fn get_tn10_config() -> configuration::Configuration {
+    configuration::Configuration {
+        base_path: "https://api-tn10.kaspa.org".to_string(),
+        user_agent: Some("OpenAPI-Generator/a6a9569/rust".to_owned()),
+        client: reqwest_middleware::ClientBuilder::new(reqwest::Client::new()).build(),
+        basic_auth: None,
+        oauth_access_token: None,
+        bearer_access_token: None,
+        api_key: None,
+    }
+}
